+++ resources/assets/js/components/RequestTrackerPanel.js
@@ -5,6 +5,7 @@ export default ( isExpanded, refreshFrequency, hasReinitPassword, isVisualizer )
 		usingReinitPassword : hasReinitPassword,
 		isVisualizer        : isVisualizer,
 		currentProfileId    : "",
+		refreshMonitor      : null,
 
 		reinitFramework(){
 			this.$refs.reinitLoader.classList.add( "cbd-spinner" );
@@ -21,6 +22,7 @@ export default ( isExpanded, refreshFrequency, hasReinitPassword, isVisualizer )
 				} );
 		},
 		loadProfilerReport( id ){
+			this.stopDebuggerMonitor();
 			fetch( `${this.appUrl}cbDebugger/renderProfilerReport`, {
 				method  : "POST",
 				headers : { "x-Requested-With": "XMLHttpRequest" },
@@ -31,13 +33,13 @@ export default ( isExpanded, refreshFrequency, hasReinitPassword, isVisualizer )
 			} )
 				.then( response => response.text() )
 				.then( html => {
-					history.pushState( { profileId: id }, null, "##" + id );
+					history.pushState( { profileId: id }, null, "#" + id );
 					this.$refs[ "cbd-profilers" ].innerHTML = html;
 					coldboxDebugger.scrollTo( "cbd-request-tracker" );
 				} );
 		},
 		clearState(){
-			history.pushState( {}, null, "##" );
+			history.pushState( {}, null, "#" );
 			this.currentProfileId = "";
 		},
 		refreshProfilers(){
@@ -65,17 +67,22 @@ export default ( isExpanded, refreshFrequency, hasReinitPassword, isVisualizer )
 				} );
 		},
 		stopDebuggerMonitor(){
-			if ( "cbdRefreshMonitor" in window ){
-				clearInterval( window.cbdRefreshMonitor );
-				console.log( "Stopped ColdBox Debugger Profiler Refresh" );
+			// stop only if loaded
+			if ( this.refreshMonitor != null ){
+				clearInterval( this.refreshMonitor );
+				this.refreshFrequency = 0;
+				console.info( "Stopped ColdBox Debugger Profiler Refresh" );
 			}
 		},
 		startDebuggerMonitor( frequency ){
-			if ( frequency == 0 ){
-				return this.stopDebuggerMonitor();
+			// Ensure monitor is stopped just in case we are switching frequencies
+			this.stopDebuggerMonitor();
+			this.refreshFrequency = frequency;
+			if ( this.refreshFrequency == 0 ){
+				return;
 			}
-			window.cbdRefreshMonitor = setInterval( this.refreshProfilers, frequency * 1000 );
-			console.log( "Started ColdBox Debugger Profiler Refresh using " + frequency + " seconds" );
+			this.refreshMonitor = setInterval( () => this.refreshProfilers(), this.refreshFrequency * 1000 );
+			console.info( "Started ColdBox Debugger Profiler Refresh using " + this.refreshFrequency + " seconds" );
 		},
 		init(){
 			window.addEventListener( "popstate", e => {